Conflicts:
	libs/batch/plugins/project-manager/src/Flowbox/ProjectManager/RPC/Handler/Graph.hs
	libs/bus/src/Flowbox/Bus/Data/Topic.hs
	libs/bus/src/Flowbox/Bus/RPC/Server/Processor.hs
	scripts/config/packages.py

+++ libs/data/graphics/src/Flowbox/Geom2D/CubicBezier/Conversion.hs
@@ -16,8 +16,8 @@ import Flowbox.Prelude
 
 
 
-fcb2gcb :: CubicBezier Double -> Cubic.CubicBezier
+fcb2gcb :: CubicBezier Float -> Cubic.CubicBezier
 fcb2gcb (CubicBezier (fp2gp -> p1) (fp2gp -> p2) (fp2gp -> p3) (fp2gp -> p4)) = Cubic.CubicBezier p1 p2 p3 p4
 
-gcb2fcb :: Cubic.CubicBezier -> CubicBezier Double
+gcb2fcb :: Cubic.CubicBezier -> CubicBezier Float
 gcb2fcb (Cubic.CubicBezier (gp2fp -> p1) (gp2fp -> p2) (gp2fp -> p3) (gp2fp -> p4)) = CubicBezier p1 p2 p3 p4
